% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{action_btn_mdl}
\alias{action_btn_mdl}
\title{Small action button}
\usage{
action_btn_mdl(...)
}
\arguments{
\item{...}{parameters same as in shiny::actionButton}
}
\value{
shiny.tag object
}
\description{
Small action button
}
\examples{
action_btn_mdl("show_how", "", icon = icon("question"))
}
