% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_sent_lab}
\alias{assign_sent_lab}
\title{Assigning labels for sentiment score}
\usage{
assign_sent_lab(x, br = c(-Inf, -0.1, 0.1, Inf))
}
\arguments{
\item{x}{numeric vector}

\item{br}{breaks for score}
}
\value{
character vector of labels
}
\description{
Assigning labels for sentiment score
}
\examples{
assign_sent_lab(3)
}
