% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{get_top_from_lda}
\alias{get_top_from_lda}
\title{Getting top topics from LDA}
\usage{
get_top_from_lda(lda_out, no_top = 2)
}
\arguments{
\item{lda_out}{object from LDA}

\item{no_top}{number of top words for topic}
}
\value{
data.table containing top word for each topic
}
\description{
Getting top topics from LDA
}
\examples{
topiced <- get_topic_reviews(review_set_example()@reviews)
get_top_from_lda(topiced)
}
