% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{get_topic_reviews}
\alias{get_topic_reviews}
\title{Getting topics from a list of reviews}
\usage{
get_topic_reviews(reviews, no_topics = 2, method = "Gibbs")
}
\arguments{
\item{reviews}{list of reviews}

\item{no_topics}{number of topics (integer)}

\item{method}{method ('Gibbs' or 'VEM')}
}
\value{
LDA_\code{method} topic model
}
\description{
Getting topics from a list of reviews
}
\examples{
get_topic_reviews(review_set_example()@reviews)
}
