% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ic_utf8}
\alias{ic_utf8}
\title{Converting string to UTF-8}
\usage{
ic_utf8(x)
}
\arguments{
\item{x}{character string}
}
\value{
x with UTF-8 encoding
}
\description{
Converting string to UTF-8
}
\examples{
ic_utf8(c("Andy", "Bob"))
}
