% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newfinn}
\alias{newfinn}
\title{Sentiment score usign affin dataset reviews}
\format{
## `newfinn`
A data table with 2477 rows and 2 columns:
\describe{
  \item{word}{word}
  \item{value}{sentiment ranges from -5 to 5}
}
}
\source{
http://www2.imm.dtu.dk/pubdb/pubs/6010-full.html
}
\usage{
newfinn
}
\description{
Sentiment score usign affin dataset reviews
}
\keyword{datasets}
