% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rating_hour}
\alias{plot_rating_hour}
\title{Title}
\usage{
plot_rating_hour(dt, lw_col = "blue", hgh_col = "red")
}
\arguments{
\item{dt}{data.table}

\item{lw_col}{color for low values}

\item{hgh_col}{color for high values}
}
\value{
\code{ggplot2} object
}
\description{
Title
}
\examples{
plot_rating_hour(sample_apps_reviews[[1]])
}
