% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modelling.R
\name{plot_top_from_lda}
\alias{plot_top_from_lda}
\title{Plot top terms from topics}
\usage{
plot_top_from_lda(ds)
}
\arguments{
\item{ds}{data.table object, result of \code{get_top_from_lda}}
}
\value{
ggplot object
}
\description{
Plot top terms from topics
}
\examples{
topiced <- get_topic_reviews(review_set_example()@reviews)
top <- get_top_from_lda(topiced)
plot_top_from_lda(top)
}
