% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_word_count}
\alias{plot_word_count}
\title{Title}
\usage{
plot_word_count(oo, mn = 10)
}
\arguments{
\item{oo}{data.table with columns 'proc_word' and 'how_many'}

\item{mn}{number indicating how many words to plot}
}
\value{
\code{ggplot2} object
}
\description{
Title
}
\examples{
set_review <- review_set(lapply(X = sample_apps_reviews$Snapchat$review[1:10], FUN = function(i) review(i)))
plot_word_count(set_review@words)
}
