% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/review.R
\name{remove_numerical_and_punct}
\alias{remove_numerical_and_punct}
\alias{remove_numerical_and_punct,character-method}
\title{Remove Numerical and Punctuation Characters}
\usage{
remove_numerical_and_punct(x)

\S4method{remove_numerical_and_punct}{character}(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A modified character vector with numerical and punctuation characters removed.
}
\description{
This method removes numerical and punctuation characters from a given character vector.
}
\examples{
remove_numerical_and_punct("Hello, 123!")
}
