% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{remove_stop_words,character-method}
\alias{remove_stop_words,character-method}
\title{Title}
\usage{
\S4method{remove_stop_words}{character}(x)
}
\arguments{
\item{character}{}
}
\value{

}
\description{
Title
}
