% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/review.R
\name{remove_stop_words}
\alias{remove_stop_words}
\alias{remove_stop_words,character-method}
\title{Removing stop words from a character}
\usage{
remove_stop_words(x)

\S4method{remove_stop_words}{character}(x)
}
\arguments{
\item{x}{character}
}
\description{
Removing stop words from a character
}
