% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revapp-package.R
\docType{package}
\name{revapp-package}
\alias{revapp}
\alias{revapp-package}
\title{revapp: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
