% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_set.R
\name{sentimentDistributionPie}
\alias{sentimentDistributionPie}
\alias{review-set-charts}
\alias{sentimentDistributionHist}
\title{Charts for set of reviews}
\usage{
sentimentDistributionPie(x, br = c(-Inf, -0.1, 0.1, Inf))

sentimentDistributionHist(x)
}
\arguments{
\item{x}{object of class \code{review-set}}

\item{br}{breaks for establish whether review is negative or positive}
}
\value{
charts
}
\description{
Charts for set of reviews
}
\examples{
review1 <- review("This is a great product.")
review2 <- review("I'm not satisfied with this product.")

set_review <- review_set(list(review1, review2))
sentimentDistributionPie(set_review)
sentimentDistributionHist(set_review)
}
