% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_set.R
\name{review-set-summary}
\alias{review-set-summary}
\alias{sentimentRatio}
\alias{sentimentSummary}
\title{Title}
\usage{
sentimentRatio(x, br = c(-Inf, -0.1, 0.1, Inf))

sentimentSummary(x)
}
\arguments{
\item{x}{object of class \code{review-set}}

\item{br}{breaks for establish whether review is negative or positive}
}
\value{
numeric
}
\description{
Title
}
\examples{
ll <- lapply(X = sample_apps_reviews$Snapchat$review[1:10],
             FUN = function(i) review(x = i))
set_review <- review_set(ll)
sentimentRatio(set_review)
sentimentSummary(set_review)
}
