% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\docType{class}
\name{review}
\alias{review}
\title{review Class}
\usage{
review(x)
}
\arguments{
\item{x}{character containing review}
}
\value{

}
\description{
A class representing a review.
}
\section{Slots}{

\describe{
\item{\code{original}}{the original text of the review.}

\item{\code{processed}}{the processed text of the review.}

\item{\code{sent_score}}{the sentiment score of the review.}
}}

\examples{
review("This is a great product.")
review <- review("This is a great product.")
review
}
