% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_set.R
\docType{class}
\name{review_set}
\alias{review_set}
\alias{review_set-class}
\title{review_set Class}
\usage{
review_set(x)
}
\arguments{
\item{x}{list of reviews}
}
\value{
object of class \code{review_set}
}
\description{
A class representing a set of reviews.
}
\section{Slots}{

\describe{
\item{\code{reviews}}{A list containing individual review objects.}

\item{\code{words}}{A data.table containing the words in the reviews.}

\item{\code{sent_scores}}{A numeric vector containing the sentiment scores of the reviews.}

\item{\code{ratings}}{A numeric vector containing ratings of the reviews}
}}

\examples{
dt_ex <- sample_apps_reviews$Snapchat
ll <- lapply(X = 1:10,
             FUN = function(i) {
                 review(x = dt_ex$review, rt = dt_ex$rating)
             })
set_review <- review_set(ll)
set_review
}
