% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_set.R
\name{review_set_example}
\alias{review_set_example}
\title{Example of 'review_set' class}
\usage{
review_set_example()
}
\value{
object of class 'review_set'
}
\description{
Example of 'review_set' class
}
\examples{
review_set_example()
}
