% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_apps_reviews}
\alias{sample_apps_reviews}
\title{Sample fo reviews}
\format{
## `sample_apps_reviews`
A data table with 2,500 rows and 8 columns:
\describe{
  \item{id}{id of an review}
  \item{review time}{time of the review}
  \item{author}{nick}
  \item{app_version}{version of the app}
  \item{title}{title of a review}
  \item{rating}{rating of an app}
  \item{review}{review itslef}
  \item{app_id}{id of a reviewed app}
}
}
\source{
appler package
}
\usage{
sample_apps_reviews
}
\description{
Sample fo reviews
}
\keyword{datasets}
