% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{sent_score}
\alias{sent_score}
\title{Calculates sentiment score for a sentence}
\usage{
sent_score(x)
}
\arguments{
\item{x}{character vector}
}
\value{
single number with mean setiment with afinn dictionairy
}
\description{
Calculates sentiment score for a sentence
}
