% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_set.R
\name{sentimentRatio}
\alias{sentimentRatio}
\title{Title}
\usage{
sentimentRatio(x, br = c(-Inf, -0.1, 0.1, Inf))
}
\arguments{
\item{br}{}
}
\value{

}
\description{
Title
}
\examples{
set_review <- review_set(lapply(X = sample_apps_reviews$Snapchat$review[1:10], FUN = function(i) review(x = i)))
sentimentRatio(set_review)
}
