% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{shiny_ui}
\alias{shiny_ui}
\alias{summary_ui_main}
\alias{summary_ui_side}
\alias{wordcloud_ui_main}
\alias{wordcloud_ui_side}
\alias{sentiment_ui_main}
\alias{sentiment_ui_side}
\alias{topics_ui_main}
\alias{topics_ui_side}
\title{UI of the app}
\usage{
summary_ui_main()

summary_ui_side()

wordcloud_ui_main()

wordcloud_ui_side()

sentiment_ui_main()

sentiment_ui_side()

topics_ui_main()

topics_ui_side()
}
\value{
\code{tabPanel} or sidePanels with UI for specific sections
}
\description{
UI of the app
}
\examples{
summary_ui_main()
}
