% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{show,review-method}
\alias{show,review-method}
\title{Print method}
\usage{
\S4method{show}{review}(object)
}
\arguments{
\item{object}{review object}
}
\value{
printed message about review
}
\description{
Print method
}
\examples{
print(review("This is a great product.", 5))
}
