% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_set.R
\name{show,review_set-method}
\alias{show,review_set-method}
\title{Title}
\usage{
\S4method{show}{review_set}(object)
}
\arguments{
\item{review_set}{object of class review}
}
\value{

}
\description{
Title
}
\examples{
ll <- lapply(X = sample_apps_reviews$Snapchat$review[1:10],
             FUN = function(i) review(x = i))
review_set(ll)
}
