% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_summary.R
\name{summary_ui}
\alias{summary_ui}
\title{Represetnation of the summary}
\usage{
summary_ui()
}
\value{
\code{tabPanel} with ui for Summary tab
}
\description{
Represetnation of the summary
}
\examples{
summary_ui()
}
