% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/review.R
\name{word_stem}
\alias{word_stem}
\alias{word_stem,character-method}
\title{Stemming words}
\usage{
word_stem(x)

\S4method{word_stem}{character}(x)
}
\arguments{
\item{x}{character}
}
\description{
Stemming words
}
