% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{word_tokenize,character-method}
\alias{word_tokenize,character-method}
\title{Tokenize words in a character string}
\usage{
\S4method{word_tokenize}{character}(x)
}
\arguments{
\item{x}{object to tokenize}
}
\value{
A character vector of individual words.
}
\description{
This method tokenizes the words in a character string by splitting it on whitespace and handling contractions.
}
\examples{
word_tokenize("I don't like it")
}
