% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{get_top_from_lda}
\alias{get_top_from_lda}
\title{Getting top topics from LDA}
\usage{
get_top_from_lda(lda_out, no_top)
}
\arguments{
\item{lda_out}{object from LDA}

\item{no_top}{number of top words for topic}
}
\value{

}
\description{
Getting top topics from LDA
}
