% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rating_hour}
\alias{plot_rating_hour}
\title{Title}
\usage{
plot_rating_hour(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
\code{ggplot2} object
}
\description{
Title
}
\examples{
plot_rating_hour(sample_apps_reviews[[1]])
}
