% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_word_count}
\alias{plot_word_count}
\title{Number of occurrences of each word}
\usage{
plot_word_count(oo, mn = 10)
}
\arguments{
\item{oo}{data.table with columns 'proc_word' and 'how_many'}

\item{mn}{number indicating how many words to plot}
}
\value{
\code{ggplot2} object
}
\description{
Number of occurrences of each word
}
\examples{
plot_word_count(review_set_example()@words)
}
