% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{pre_process}
\alias{pre_process}
\alias{pre_process,review-method}
\title{Preprocess character}
\usage{
\S4method{pre_process}{character}(object)

\S4method{pre_process}{review}(object)
}
\arguments{
\item{object}{An object to preprocess}
}
\value{
returns processed character, that means after removing emojis, converting to lowercase,
removing numeric values and punctuations, tokenizing, removing stop words and stemming
}
\description{
Preprocess character
}
