% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revapp-package.R
\docType{package}
\name{revapp-package}
\alias{revapp}
\alias{revapp-package}
\title{revapp: Review Appstore}
\description{
Analysis of sentiment in AppStore Using text mining methods
}
\author{
\strong{Maintainer}: Dominik Zabinski \email{dzabinski@studentu.uw.edu.pl}

Authors:
\itemize{
  \item Palina Tkachova \email{ptkachova@studentu.uw.edu.pl}
}

}
\keyword{internal}
