% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_set.R
\name{sentimentDistributionPie}
\alias{sentimentDistributionPie}
\alias{review-set-summary}
\alias{sentimentDistributionHist}
\alias{sentimentRatio}
\alias{sentimentSummary}
\title{Set of functions for summary of 'review_set' class in terms of sentiment}
\usage{
sentimentDistributionPie(
  x,
  br = c(-Inf, -0.1, 0.1, Inf),
  colors = c("blue", "red", "green", "white")
)

sentimentDistributionHist(x)

sentimentRatio(x, br = c(-Inf, -0.1, 0.1, Inf))

sentimentSummary(x)
}
\arguments{
\item{x}{object of class \code{review-set}}

\item{br}{breaks for establish whether review is negative or positive}

\item{colors}{colors for categories}
}
\value{
charts
}
\description{
Set of functions for summary of 'review_set' class in terms of sentiment
}
\examples{
sentimentDistributionPie(review_set_example())
sentimentDistributionHist(review_set_example())
sentimentRatio(review_set_example())
sentimentSummary(review_set_example())
}
