% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\docType{class}
\name{review}
\alias{review}
\alias{review-class}
\title{review Class}
\usage{
review(x, rt)
}
\arguments{
\item{x}{character containing review}

\item{rt}{rating of a review}
}
\value{
object of class \code{review}
}
\description{
A class representing a review.
}
\section{Slots}{

\describe{
\item{\code{original}}{the original text of the review.}

\item{\code{processed}}{the processed text of the review.}

\item{\code{sent_score}}{the sentiment score of the review.}

\item{\code{rating}}{rating assigned to the review}
}}

\examples{
review <- review("This is a great product.", 5)
review
}
