% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\docType{class}
\name{review}
\alias{review}
\title{review Class}
\usage{
review(x)
}
\arguments{
\item{x}{character containing review}
}
\value{

}
\description{
A class representing a review.
}
\section{Slots}{

\describe{
\item{\code{original}}{The original text of the review.}

\item{\code{processed}}{The processed text of the review.}

\item{\code{sent_score}}{The sentiment score of the review.}
}}

\examples{
review <- review("This is a great product.")
review
review <- review("This is a great product.")
review
}
