% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_set.R
\docType{class}
\name{review_set}
\alias{review_set}
\title{review_set Class}
\usage{
review_set(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
A class representing a set of reviews.
}
\section{Slots}{

\describe{
\item{\code{reviews}}{A list containing individual review objects.}

\item{\code{words}}{A data.table containing the words in the reviews.}

\item{\code{sent_scores}}{A numeric vector containing the sentiment scores of the reviews.}
}}

\examples{
review1 <- review("This is a great product.")
review2 <- review("I'm not satisfied with this product.")

review_set <- review_set(list(review1, review2))
review_set
}
