% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{sent_score}
\alias{sent_score}
\title{Calculates sentiment score for a sentence}
\source{
http://www2.imm.dtu.dk/pubdb/pubs/6010-full.html
}
\usage{
sent_score(x)
}
\arguments{
\item{x}{character vector}
}
\value{
single number with mean sentiment with afinn dictionary
}
\description{
Calculates sentiment score for a sentence
}
\examples{
sent_score(c("great", "product"))
}
