% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_set.R
\name{show,review_set-method}
\alias{show,review_set-method}
\title{Show method for 'review_set'}
\usage{
\S4method{show}{review_set}(object)
}
\arguments{
\item{object}{object of class review}
}
\value{
print of \code{review_set} object
}
\description{
Show method for 'review_set'
}
\examples{
print(review_set_example())
}
